controllers

See merge request projetBDI-GR06/serveur!9

+++ src/main/java/fr/miage/gromed/dto/PresentationFicheDto.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Getter
 @Setter
 public class PresentationFicheDto extends PresentationDto implements Serializable {
-    private final double stock;
+    private final int stock;
     private final String statutAdmin;
